# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for Live-information regarding use of Chalmers facilities",
    "version": "0.0.1",
    "title": "Liveinfo Chalmers"
  },
  "host": "liveinfo.chalmers.se",
  "basePath": "/api",
  "tags": [],
  "schemes": [
    "https",
    "ws"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/rooms": {
      "get": {
        "tags": [
          "Room information"
        ],
        "summary": "Receive room information",
        "description": ".",
        "operationId": "getRooms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "query",
            "description": "access key for api",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/roomsList"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Add Rooms"
        ],
        "summary": "Add rooms",
        "description": ".",
        "operationId": "addRooms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rooms",
            "in": "body",
            "description": "Room added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addRoomDef"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "access key for api",
            "required": true,
            "type": "string"
          }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
              "$ref": "#/definitions/addRoomDef"
            }
          }
        }
      }
    },
    "/rooms/{buildingID}": {
      "get": {
        "tags": [
          "Building ID"
        ],
        "summary": "Receive Building ID",
        "description": ".",
        "operationId": "getBuildingID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "buildingID",
            "in": "path",
            "description": "Rooms in a building",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/buildingId"
            }
          }
        }
      }
    },
    "/rooms/occupants/history/{roomId}":{
      "get": {
        "tags": [
          "Room history"
        ],
        "summary": "Get history of a room",
        "description": ".",
        "operationId": "getRoomHistory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "description": "History of room",
            "required": true,
            "type": "string"
          }
        ],
      "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/roomId"
            }
          }
        }
      }
    },
    "/rooms/status":{
      "get": {
        "tags": [
          "Room status"
        ],
        "summary": "Get status of all rooms",
        "description": ".",
        "operationId": "getRoomStatus",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/roomStatus"
            }
          }
        }
      }
    },
    "/sensors": {
      "post": {
        "tags": [
          "Add Sensors"
        ],
        "summary": "Add sensors",
        "description": "Add a new sensor to the system.",
        "operationId": "addSensors",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sensor",
            "in": "body",
            "description": "Sensor details including sensorId and roomId",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addSensorsDesc"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Access key for API",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor added successfully",
            "schema": {
              "$ref": "#/definitions/addSensorsDesc"
            }
          }
        }
      }
    },
    "/sensors/report": {
      "post": {
        "tags": [
          "Sensors report"
        ],
        "summary": "Submit sensor data",
        "description": "Report sensor data including firmware version, sensorId, and sensor states.",
        "operationId": "sensorData",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sensor report",
            "in": "body",
            "description": "Sensor reporting data including firmwareVersion, sensorId, and other details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sensorReport"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Access key for API",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Data reported successfully",
            "schema": {
              "$ref": "#/definitions/sensorReport"
            }
          }
        }
      }
    },
    "/sensors/report/test": {
      "get": {
        "tags": [
          "Report test data"
        ],
        "summary": "Test data from sensors",
        "description": ".",
        "operationId": "testData",
        "produces": [
          "application/json"
        ],
        "parameters": [
           {
            "name": "authorization",
            "in": "header",
            "description": "access key for api",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/sensorTest"
            }
          }
        }
      }
    },
    "/sensors/dead": {
      "get": {
        "tags": [
          "Get dead sensors"
        ],
        "summary": "Get all dead sensors",
        "description": ".",
        "operationId": "getDeadSensors",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/deadSensors"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "roomsList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "state": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "buildingId",
          "name",
          "description",
          "state"
        ]
      }
    },
    "buildingId": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      }
    },
    "roomId": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "occupants": {
            "type": "string"
          }
        },
        "required": ["occupants"]
      }
    },
    "roomStatus": {
     "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "state": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "buildingId",
          "name",
          "description",
          "state"
        ]
      }
    },
    "addSensorsDesc": {
      "type": "object",
        "properties": {
          "sensorId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          }
        },
        "required": [
          "sensorId",
          "roomId"
        ]
    },
    "addRoomDef": {
      "type": "object",
          "properties": {
            "building": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "coordinates": {
              "type": "string"
            }
          },
          "required": [
            "building",
            "name",
            "description",
            "coordinates"
        ]
    },
    "sensorTest": {
      "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "firmwareVersion": {
              "type": "string"
            },
            "sensorId": {
              "type": "string"
            },
            "roomName": {
              "type": "string"
            },
            "occupants": {
              "type": "integer"
            },
            "radarState": {
              "type": "integer"
            },
            "pirState": {
              "type": "boolean"
            }
          },
          "required": [
            "firmwareVersion",
            "sensorId",
            "roomName",
            "occupants",
            "radarState",
            "pirState"
        ]
      }
    },
    "deadSensors": {
      "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "firmwareVersion": {
              "type": "string"
            },
            "sensorId": {
              "type": "string"
            },
            "roomName": {
              "type": "string"
            },
            "occupants": {
              "type": "integer"
            },
            "radarState": {
              "type": "integer"
            },
            "pirState": {
              "type": "integer"
            }
          },
          "required": [
            "firmwareVersion",
            "sensorId",
            "roomName",
            "occupants",
            "radarState",
            "pirState"
        ]
      }
    },
    "sensorReport": {
      "type": "object",
          "properties": {
            "firmwareVersion": {
              "type": "string"
            },
            "sensorId": {
              "type": "string"
            },
            "occupants": {
              "type": "integer"
            },
            "radarState": {
              "type": "integer"
            },
            "pirState": {
              "type": "integer"
            }
          },
          "required": [
            "firmwareVersion",
            "sensorId",
            "occupants",
            "radarState",
            "pirState"
        ]
      }
    }
}